requirements.txt
# RCA Engine Requirements
# Production-grade components

# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database (choose based on your database)
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.1.0  # MySQL
# sqlalchemy>=2.0.0  # SQL abstraction (optional)

# LLM Provider
openai>=1.0.0  # OpenAI API

# Authentication
PyJWT>=2.8.0  # JWT tokens

# Caching
redis>=5.0.0  # Redis cache (optional, memory cache available)

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Optional: Monitoring
# prometheus-client>=0.18.0

# Optional: Structured logging
# structlog>=23.1.0

